python/backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai>=0.8.0
qdrant-client>=1.11.0
sentence-transformers>=2.2.0
PyPDF2>=3.0.0
